Add a router and route mapping file
Previously, we added various caching metadata in the services and
passed those to the Client so that the Client knew where to cache the
data. This was a little messy and decentralized the offline support
code.

By adding a router, we can match against the path the Client receives
and determine where to cache the data.